[Refactor] Move types into base
