+++ be/src/column/decimalv3_column.cpp
@@ -14,7 +14,7 @@
 
 #include "column/decimalv3_column.h"
 
-#include "base/int256.h"
+#include "base/types/int256.h"
 #include "column/fixed_length_column.h"
 
 namespace starrocks {
