+++ be/src/column/fixed_length_column_base.cpp
@@ -14,8 +14,9 @@
 
 #include "column/fixed_length_column_base.h"
 
-#include "base/int256.h"
 #include "base/simd/gather.h"
+#include "base/types/decimal12.h"
+#include "base/types/int256.h"
 #include "column/column_helper.h"
 #include "column/vectorized_fwd.h"
 #include "common/config.h"
@@ -23,7 +24,6 @@
 #include "gutil/casts.h"
 #include "gutil/strings/fastmem.h"
 #include "gutil/strings/substitute.h"
-#include "storage/decimal12.h"
 #include "types/large_int_value.h"
 #include "util/hash_util.hpp"
 #include "util/mysql_row_buffer.h"
