+++ be/src/formats/parquet/encoding_plain.h
@@ -20,9 +20,9 @@
 #include "base/bit/bit_util.h"
 #include "base/coding.h"
 #include "base/container/raw_container.h"
-#include "base/int256.h"
 #include "base/string/faststring.h"
 #include "base/string/slice.h"
+#include "base/types/int256.h"
 #include "column/column.h"
 #include "column/column_helper.h"
 #include "column/nullable_column.h"
