+++ be/src/formats/parquet/statistics_helper.cpp
@@ -17,6 +17,7 @@
 #include <string>
 
 #include "base/simd/simd.h"
+#include "base/types/uint24.h"
 #include "column/column_helper.h"
 #include "column/type_traits.h"
 #include "column/vectorized_fwd.h"
@@ -31,7 +32,6 @@
 #include "runtime/types.h"
 #include "storage/column_predicate.h"
 #include "storage/types.h"
-#include "storage/uint24.h"
 #include "types/date_value.h"
 #include "types/large_int_value.h"
 #include "types/logical_type.h"
