+++ be/src/storage/type_traits.h
@@ -36,8 +36,8 @@
 
 #include <cstdint>
 
-#include "base/int256.h"
 #include "base/string/slice.h"
+#include "base/types/int256.h"
 #include "runtime/decimalv2_value.h"
 #include "storage/collection.h"
 #include "storage/olap_common.h"
