+++ be/src/storage/types.cpp
@@ -37,6 +37,7 @@
 #include "base/hash/unaligned_access.h"
 #include "base/string/slice.h"
 #include "base/string/string_parser.hpp"
+#include "base/types/decimal12.h"
 #include "base/utility/mem_util.hpp"
 #include "base/utility/mysql_global.h"
 #include "gutil/strings/numbers.h"
@@ -47,7 +48,6 @@
 #include "runtime/types.h"
 #include "storage/collection.h"
 #include "storage/convert_helper.h"
-#include "storage/decimal12.h"
 #include "storage/decimal_type_info.h"
 #include "storage/olap_define.h"
 #include "storage/olap_type_infra.h"
