+++ be/src/util/date_func.h
@@ -21,7 +21,7 @@
 #include <ctime>
 #include <string>
 
-#include "storage/uint24.h"
+#include "base/types/uint24.h"
 
 namespace starrocks {
