+++ be/src/util/mysql_row_buffer.cpp
@@ -41,7 +41,7 @@
 #include <cstdio>
 #include <type_traits>
 
-#include "base/int256.h"
+#include "base/types/int256.h"
 #include "base/utility/mysql_global.h"
 #include "common/logging.h"
 #include "gutil/strings/fastmem.h"
