+++ be/test/base/int256_test.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "base/int256.h"
+#include "base/types/int256.h"
 
 #include <gtest/gtest.h>
